# docs/site/public/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditPage": {
        "description": "A page of the trail, newest first, with the cursor for the next one.",
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "group_id": {
            "title": "Group Id",
            "type": "integer"
          },
          "next_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "pass as `before` for the following page; null on the last one",
            "title": "Next Before"
          }
        },
        "required": [
          "group_id",
          "events"
        ],
        "title": "AuditPage",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "FastAPI's error envelope \u2014 `{\"detail\": \"...\"}` \u2014 named so the schema can\npoint at it and a generated client can read it.",
        "properties": {
          "detail": {
            "description": "a sentence for a developer, never for an end user",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "The HTTP surface behind the Telegram Mini App and the web console.",
    "title": "Cookiebot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/groups/{group_id}/audit": {
      "get": {
        "description": "The group's trail, newest first, keyset-paginated (D11 \u2014 no OFFSET, no\nunbounded list).\n\n`next_before` is the cursor for the following page and is `null` on the last\none. Filtering by `action` or `actor_user_id` narrows without changing the\ncursor's meaning.",
        "operationId": "read_audit_groups__group_id__audit_get",
        "parameters": [
          {
            "description": "the Telegram chat id \u2014 negative, as Telegram writes it",
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "description": "the Telegram chat id \u2014 negative, as Telegram writes it",
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "description": "events per page",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "events per page",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "last id of the previous page",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "last id of the previous page",
              "title": "Before"
            }
          },
          {
            "description": "only this action, e.g. `config.updated`",
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "only this action, e.g. `config.updated`",
              "title": "Action"
            }
          },
          {
            "description": "only changes this Telegram user made",
            "in": "query",
            "name": "actor_user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "only changes this Telegram user made",
              "title": "Actor User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no bearer token, or one that did not verify"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "an admin whose token lacks the scope; ask /oauth2/token for a better one"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no such group \u2014 or the caller does not administer it, which answers alike"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read the group's audit trail, newest first",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{group_id}/config": {
      "get": {
        "description": "The group's effective settings \u2014 stored values over tenant defaults over\nv1's defaults, which is the same resolution the bot itself reads.",
        "operationId": "read_config_groups__group_id__config_get",
        "parameters": [
          {
            "description": "the Telegram chat id \u2014 negative, as Telegram writes it",
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "description": "the Telegram chat id \u2014 negative, as Telegram writes it",
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no bearer token, or one that did not verify"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "an admin whose token lacks the scope; ask /oauth2/token for a better one"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no such group \u2014 or the caller does not administer it, which answers alike"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read a group's settings",
        "tags": [
          "groups"
        ]
      },
      "patch": {
        "description": "Change some settings. Absent fields are left alone \u2014 this is a PATCH,\nnot a PUT, because a Mini App form that round-trips every column would\noverwrite a change another admin made while it was open.\n\nAn empty patch is a 400 rather than a no-op success: a client that sent\nnothing meant to send something.",
        "operationId": "update_config_groups__group_id__config_patch",
        "parameters": [
          {
            "description": "the Telegram chat id \u2014 negative, as Telegram writes it",
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "description": "the Telegram chat id \u2014 negative, as Telegram writes it",
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "the patch carried no settings"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no bearer token, or one that did not verify"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "an admin whose token lacks the scope; ask /oauth2/token for a better one"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no such group \u2014 or the caller does not administer it, which answers alike"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change some of a group's settings",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/login": {
      "post": {
        "description": "Exchange a Telegram login-widget payload for a JWT.\n\nv1 `Server.py:59-76`. The status codes and the two error bodies are v1's,\nbecause the client branches on them.",
        "operationId": "login_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginToken"
                }
              }
            },
            "description": "v1's body, key for key"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginError"
                }
              }
            },
            "description": "an empty payload \u2014 v1's `Missing data`"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginError"
                }
              }
            },
            "description": "the signature did not verify, or the payload is stale where `CB_WEBHUB_AUTH_MAX_AGE_SECONDS` is set \u2014 one answer for both, so a forgery cannot learn which of the two it got closest to"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange a Telegram login-widget payload for a JWT (v1's endpoint)",
        "tags": [
          "webhub"
        ]
      }
    },
    "/me": {
      "get": {
        "description": "Who the token says you are, what it lets you do, and which groups you\nadminister.\n\nThe Mini App's first call: it has `initData` telling it who the user is,\nbut not which of that user's groups this deployment knows about, nor\nwhether they run the whole thing. The list comes from `group_admins`,\nwhich the gateway maintains \u2014 so a promotion\nmade a minute ago appears once the bot has seen an admin-gated command in\nthat group, and not before (`cb_api.security`'s note on why this service\nnever calls Telegram to refresh it).",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no bearer token, or one that did not verify"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The caller, their scopes, and the groups they administer",
        "tags": [
          "groups"
        ]
      }
    },
...
  },
...
}